backend/pyproject.toml
[project]
name = "game-theory-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.9.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-google-genai>=2.1.12",
    "langchain-pinecone>=0.2.12",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.1.1",
    "pypdf>=6.1.3",
]
